Misc fixes in remote support dialog
