+++ src/app/utils/fp.js
@@ -190,12 +190,9 @@ export const updateInArray = curry((predicateFn, updateFn, arr) =>
 
 // Like `updateInArray` but stops after finding the element to update
 // Also like ramda `adjust` but using a predicateFn
-export const adjustWith = curry((predicateFn, updateFn, arr) => {
-  console.log(findIndex(predicateFn, arr), arr)
-  const adjusted = adjust(findIndex(predicateFn, arr), updateFn, arr)
-  console.log('adjusted', adjusted)
-  return adjusted
-})
+export const adjustWith = curry((predicateFn, updateFn, arr) =>
+  adjust(findIndex(predicateFn, arr), updateFn, arr)
+)
 
 // Like ramda `update` but using a predicateFn
 export const updateWith = curry((predicateFn, newValue, arr) =>
